move AdvSSScreenshotObj and DisplayMessage()
